Factor server/ to match litestar-fullstack pattern
Move implementations out of server/core.py into lib/:
- Exception handlers → lib/exceptions.py
- Dependency providers → lib/deps.py
- Sentry init → lib/sentry.py (new)

Add server/config.py as the config layer between settings and plugins:
settings.py → config.py (framework configs) → plugins.py (plugin instances)

server/core.py is now pure wiring: ApplicationCore(InitPluginProtocol)
references lib/ for implementations and server/ for config/plugins.
