+++ scripts/create_all_app_metadata.py
@@ -2,6 +2,6 @@ if __name__ == "__main__":
     import asyncio
 
     from convergence_games.server.app import app
-    from convergence_games.server.plugins import sqlalchemy_config as config
+    from convergence_games.server.config import sqlalchemy as config
 
     asyncio.run(config.create_all_metadata(app))
